fix: change OAuth column type

+++ auth/src/main/kotlin/cobo/auth/data/entity/Oauth.kt
@@ -16,7 +16,7 @@ data class Oauth(
     var id: Int?,
     @ManyToOne
     var user: User,
-    var oauthId: Int,
+    var oauthId: Long,
     @Enumerated(EnumType.ORDINAL)
     var oauthType: OauthTypeEnum,
     var accessToken: String?
